refactor: 나의 trip 조회 쿼리 메서드 trip 패키지로 이동
